requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.27
pydantic==2.6.3
bcrypt==4.1.2
PyJWT==2.8.0
requests==2.31.0